Completed stretch challenge and fixed cart link
